pac CLI expects BAP regions/language names, but existing pipelines use e.g. language = 'English', but BAP expects 'English (United States)'
